refactor : backend 폴더에 적용 됐던 괄호 제외
