+++ app/api/user/login/route.ts
@@ -1,6 +1,6 @@
-import { LoginRequestDTO } from '@/(backend)/uesrs/auths/login/applications/dtos/LoginRequest';
-import { LoginUseCase } from '@/(backend)/uesrs/auths/login/applications/usecases/LoginUseCase';
-import { LoginRepository } from '@/(backend)/uesrs/auths/login/infrastructures/LoginRepository';
+import { LoginRequestDTO } from '@/backend/uesrs/auths/login/applications/dtos/LoginRequest';
+import { LoginUseCase } from '@/backend/uesrs/auths/login/applications/usecases/LoginUseCase';
+import { LoginRepository } from '@/backend/uesrs/auths/login/infrastructures/LoginRepository';
 import { cookies } from 'next/headers';
 import { NextRequest, NextResponse } from 'next/server';
