+++ backend/juniors/ranks/domains/entities/JuniorRank.ts
