+++ backend/uesrs/auths/login/domains/entities/LoginUserEntity.ts
