+++ backend/uesrs/auths/login/domains/repository/LoginRepositoryInterface.ts
