+++ backend/uesrs/domains/entities/CommonAuthEntity.ts
